+++ ChangeLog.rst
@@ -133,3 +133,4 @@
 * Fix #39: Documentation update
 * Fix #61: Functionality added for optional use of automatic links
 * Feature #80: 'title' attribute is preserved in both inline and reference links
+* Feature #82: More command line options. See docs.